# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EscapeHint API",
    "version": "1.0.2",
    "description": "API specification for the EscapeHint service, providing real-time hints for escape room players and management tools for administrators. Updated to reflect actual backend implementation."
  },
  "servers": [
    {
      "url": "https://escapehint.vercel.app/api",
      "description": "Production Server"
    }
  ],
...
  "paths": {
    "/themes": {
      "get": {
        "tags": ["Player"],
        "summary": "Get Playable Themes",
        "description": "Retrieves a list of themes available for players. By default, it returns only active themes.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter themes by status. Only 'active' is allowed for players.",
            "schema": {
              "type": "string",
              "enum": ["active"],
              "default": "active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of playable themes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Theme" }
                }
              }
            }
          }
        }
      }
    },
    "/themes/{themeId}": {
      "get": {
        "tags": ["Player"],
        "summary": "Get Theme Details",
        "description": "Retrieves detailed information about a specific theme by its ID.",
        "parameters": [{ "$ref": "#/components/parameters/ThemeId" }],
        "responses": {
          "200": {
            "description": "Theme details retrieved successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Theme" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/admin/themes": {
      "get": {
        "tags": ["Admin-Themes"],
        "summary": "Get All Themes (Admin)",
        "description": "Retrieves a list of all themes, including active and inactive ones. Includes hint count for UI display.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "A list of all themes with their hint counts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Theme" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Admin-Themes"],
        "summary": "Create a New Theme",
        "description": "Adds a new escape room theme to the system.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThemeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Theme created successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Theme" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/admin/themes/{themeId}": {
      "put": {
        "tags": ["Admin-Themes"],
        "summary": "Update a Theme",
        "description": "Updates the details of an existing theme (name, description, etc.).",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ThemeId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThemeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Theme updated successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Theme" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Admin-Themes"],
        "summary": "Delete a Theme",
        "description": "Deletes a theme and all its associated hints (cascade delete). Cannot be deleted if an associated game session is currently in progress.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ThemeId" }],
        "responses": {
          "204": { "description": "Theme deleted successfully." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Theme": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "readOnly": true },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "playTime": { "type": "integer", "description": "in minutes" },
          "isActive": { "type": "boolean" },
          "difficulty": {
            "type": "string",
            "description": "Visual representation of difficulty, e.g., '★★★☆☆'",
            "nullable": true
          },
          "hintCount": {
            "type": "integer",
            "description": "Number of hints in this theme.",
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
      "CreateThemeRequest": {
        "type": "object",
        "required": ["name", "playTime"],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "example": "Zombie Lab"
          },
          "description": { "type": "string", "nullable": true },
          "playTime": {
            "type": "integer",
            "minimum": 10,
            "maximum": 180,
            "example": 60
          },
          "isActive": { "type": "boolean", "default": true },
          "difficulty": {
            "type": "string",
            "nullable": true,
            "example": "★★★★☆"
          }
        }
      },
      "UpdateThemeRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 50 },
          "description": { "type": "string", "nullable": true },
          "playTime": { "type": "integer", "minimum": 10, "maximum": 180 },
          "isActive": { "type": "boolean" },
          "difficulty": { "type": "string", "nullable": true }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" }
        }
      }
    },
    "parameters": {
...
      "ThemeId": {
        "name": "themeId",
        "in": "path",
        "required": true,
        "description": "Unique identifier of the theme.",
        "schema": { "type": "string", "format": "uuid" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - The server cannot or will not process the request due to something that is perceived to be a client error (e.g., malformed request syntax).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - The client must authenticate itself to get the requested response. JWT token is missing or invalid.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "NotFound": {
        "description": "Not Found - The server can't find the requested resource.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Conflict": {
        "description": "Conflict - The request conflicts with the current state of the server (e.g., resource already exists, or a theme with active sessions cannot be deleted).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token in 'Value' field: Bearer <JWT>"
      }
    }
  }
}